refactor: 티켓 모달 컴포넌트 수정 및 입금 링크 컴포넌트 추가
- 입금링크와 기존 QR 코드를 분기처리하여 표시하는 컴포넌트를 추가했습니다.(QrcodeAndDeposit.tsx)
- 기존 티켓 모달 컴포넌트에서 getQueryData로 선택한 티켓의 정보를 가져오는 코드를 props로 변경하였습니다.
